moved function to parse a list of strings from TriggerProcessor to TextUtils
