+++ trunk/workflow-manager/src/main/java/org/mitre/mpf/wfm/util/TextUtils.java
@@ -123,4 +123,43 @@ public class TextUtils {
 			return a.compareTo(b);
 		}
 	}
+
+    public static List<String> parseListFromString(String listOfStrings) {
+        if (!listOfStrings.contains(";") && !listOfStrings.contains("\\")) {
+            return Arrays.asList(listOfStrings.strip());
+        }
+
+        var values = new ArrayList<String>();
+        var currentSegment = new StringBuilder();
+        boolean inEscapeSequence = false;
+        for (int i = 0; i < listOfStrings.length(); i++) {
+            char ch = listOfStrings.charAt(i);
+            if (inEscapeSequence) {
+                inEscapeSequence = false;
+                currentSegment.append(ch);
+                continue;
+            }
+            switch (ch) {
+                case '\\' -> inEscapeSequence = true;
+                case ';' -> {
+                    var newValue = currentSegment.toString().strip();
+                    if (!newValue.isBlank()) {
+                        values.add(newValue);
+                    }
+                    currentSegment.setLength(0);
+                }
+                default -> currentSegment.append(ch);
+            }
+        }
+
+        if (!currentSegment.isEmpty()) {
+            var newValue = currentSegment.toString().strip();
+            if (!newValue.isBlank()) {
+                values.add(newValue);
+            }
+        }
+        return values;
+    }
+
+
 }
